refactor(login): move to ssr routes
