+++ src/backend/src/auth/mod.rs
@@ -1,2 +1,1 @@
 pub mod auth_middleware;
-pub mod login;
