+++ src/backend/src/ssr_routes/mod.rs
@@ -1,0 +1,6 @@
+pub mod login;
+pub mod post_login;
+
+// test modules
+#[cfg(test)]
+pub mod tests;
