+++ src/backend/src/ssr_routes/tests/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+pub mod login_test;
